+++ packages/wrangler/src/__tests__/worker-namespace.test.ts
@@ -1,6 +1,7 @@
+import { rest } from "msw";
 import { mockAccountId, mockApiToken } from "./helpers/mock-account-id";
-import { setMockResponse, unsetAllMocks } from "./helpers/mock-cfetch";
 import { mockConsoleMethods } from "./helpers/mock-console";
+import { msw } from "./helpers/msw";
 import { runInTempDir } from "./helpers/run-in-tmp";
 import { runWrangler } from "./helpers/run-wrangler";
 
@@ -9,10 +10,6 @@ describe("dispatch-namespace", () => {
 	const std = mockConsoleMethods();
 	mockAccountId();
 	mockApiToken();
-
-	afterEach(() => {
-		unsetAllMocks();
-	});
 
 	it("should should display a list of available subcommands, for dispatch-namespace with no subcommand", async () => {
 		await runWrangler("dispatch-namespace");
@@ -45,31 +42,30 @@ describe("dispatch-namespace", () => {
 	});
 
 	describe("create namespace", () => {
-		function mockCreateRequest(expectedName: string) {
-			const requests = { count: 0 };
-			setMockResponse(
-				"/accounts/:accountId/workers/dispatch/namespaces",
-				([_url], init) => {
-					requests.count += 1;
-
-					const incomingText = init.body?.toString() || "";
-					const { name: namespace_name } = JSON.parse(incomingText);
-
-					expect(init.method).toBe("POST");
-					expect(namespace_name).toEqual(expectedName);
-
-					return {
-						namespace_id: "some-namespace-id",
-						namespace_name: "namespace-name",
-						created_on: "2022-06-29T14:30:08.16152Z",
-						created_by: "1fc1df98cc4420fe00367c3ab68c1639",
-						modified_on: "2022-06-29T14:30:08.16152Z",
-						modified_by: "1fc1df98cc4420fe00367c3ab68c1639",
-					};
+		const namespaceName = "my-namespace";
+		let counter = 0;
+		msw.use(
+			rest.post(
+				"/accounts/:accountId/workers/dispatch/namespaces/:namespaceNameParam",
+				(req, res, cxt) => {
+					counter++;
+					const { namespaceNameParam } = req.params;
+					expect(counter).toBe(1);
+					expect(namespaceNameParam).toBe(namespaceName);
+					return res.once(
+						cxt.status(200),
+						cxt.json({
+							namespace_id: "some-namespace-id",
+							namespace_name: "namespace-name",
+							created_on: "2022-06-29T14:30:08.16152Z",
+							created_by: "1fc1df98cc4420fe00367c3ab68c1639",
+							modified_on: "2022-06-29T14:30:08.16152Z",
+							modified_by: "1fc1df98cc4420fe00367c3ab68c1639",
+						})
+					);
 				}
-			);
-			return requests;
-		}
+			)
+		);
 
 		it("should display help for create", async () => {
 			await expect(
@@ -95,10 +91,7 @@ describe("dispatch-namespace", () => {
 		});
 
 		it("should attempt to create the given namespace", async () => {
-			const namespaceName = "my-namespace";
-			const requests = mockCreateRequest(namespaceName);
 			await runWrangler(`dispatch-namespace create ${namespaceName}`);
-			expect(requests.count).toEqual(1);
 
 			expect(std.out).toMatchInlineSnapshot(
 				`"Created dispatch namespace \\"my-namespace\\" with ID \\"some-namespace-id\\""`
@@ -107,21 +100,20 @@ describe("dispatch-namespace", () => {
 	});
 
 	describe("delete namespace", () => {
-		function mockDeleteRequest(expectedName: string) {
-			const requests = { count: 0 };
-			setMockResponse(
-				"/accounts/:accountId/workers/dispatch/namespaces/:namespaceName",
-				([_url, _, namespaceName], init) => {
-					requests.count += 1;
-
-					expect(init.method).toBe("DELETE");
-					expect(namespaceName).toEqual(expectedName);
-
-					return null;
+		const namespaceName = "my-namespace";
+		let counter = 0;
+		msw.use(
+			rest.delete(
+				"/accounts/:accountId/workers/dispatch/namespaces/:namespaceNameParam",
+				(req, res, cxt) => {
+					counter++;
+					const { namespaceNameParam } = req.params;
+					expect(counter).toBe(1);
+					expect(namespaceNameParam).toBe(namespaceName);
+					return res.once(cxt.status(200), cxt.json(null));
 				}
-			);
-			return requests;
-		}
+			)
+		);
 
 		it("should display help for delete", async () => {
 			await expect(
@@ -147,10 +139,7 @@ describe("dispatch-namespace", () => {
 		});
 
 		it("should try to delete the given namespace", async () => {
-			const namespaceName = "my-namespace";
-			const requests = mockDeleteRequest(namespaceName);
 			await runWrangler(`dispatch-namespace delete ${namespaceName}`);
-			expect(requests.count).toBe(1);
 
 			expect(std.out).toMatchInlineSnapshot(
 				`"Deleted dispatch namespace \\"my-namespace\\""`
@@ -159,27 +148,30 @@ describe("dispatch-namespace", () => {
 	});
 
 	describe("get namespace", () => {
-		function mockInfoRequest(expectedName: string) {
-			const requests = { count: 0 };
-			setMockResponse(
-				"/accounts/:accountId/workers/dispatch/namespaces/:namespaceName",
-				([_url, _, namespaceName], _init) => {
-					requests.count += 1;
-
-					expect(namespaceName).toEqual(expectedName);
-
-					return {
-						namespace_id: "some-namespace-id",
-						namespace_name: "namespace-name",
-						created_on: "2022-06-29T14:30:08.16152Z",
-						created_by: "1fc1df98cc4420fe00367c3ab68c1639",
-						modified_on: "2022-06-29T14:30:08.16152Z",
-						modified_by: "1fc1df98cc4420fe00367c3ab68c1639",
-					};
+		const namespaceName = "my-namespace";
+		let counter = 0;
+		msw.use(
+			rest.get(
+				"/accounts/:accountId/workers/dispatch/namespaces/:namespaceNameParam",
+				(req, res, cxt) => {
+					counter++;
+					const { namespaceNameParam } = req.params;
+					expect(counter).toBe(1);
+					expect(namespaceNameParam).toBe(namespaceName);
+					return res.once(
+						cxt.status(200),
+						cxt.json({
+							namespace_id: "some-namespace-id",
+							namespace_name: "namespace-name",
+							created_on: "2022-06-29T14:30:08.16152Z",
+							created_by: "1fc1df98cc4420fe00367c3ab68c1639",
+							modified_on: "2022-06-29T14:30:08.16152Z",
+							modified_by: "1fc1df98cc4420fe00367c3ab68c1639",
+						})
+					);
 				}
-			);
-			return requests;
-		}
+			)
+		);
 
 		it("should display help for get", async () => {
 			await expect(
@@ -205,10 +197,7 @@ describe("dispatch-namespace", () => {
 		});
 
 		it("should attempt to get info for the given namespace", async () => {
-			const namespaceName = "my-namespace";
-			const requests = mockInfoRequest(namespaceName);
 			await runWrangler(`dispatch-namespace get ${namespaceName}`);
-			expect(requests.count).toBe(1);
 
 			expect(std.out).toMatchInlineSnapshot(`
 			"{
@@ -224,32 +213,33 @@ describe("dispatch-namespace", () => {
 	});
 
 	describe("list namespaces", () => {
-		function mockListRequest() {
-			const requests = { count: 0 };
-			setMockResponse(
-				"/accounts/:accountId/workers/dispatch/namespaces",
-				([_url, _, _page, _perPage], _init) => {
-					requests.count += 1;
-
-					return [
-						{
+		const namespaceName = "my-namespace";
+		let counter = 0;
+		msw.use(
+			rest.get(
+				"/accounts/:accountId/workers/dispatch/namespaces/:namespaceNameParam",
+				(req, res, cxt) => {
+					counter++;
+					const { namespaceNameParam } = req.params;
+					expect(counter).toBe(1);
+					expect(namespaceNameParam).toBe(namespaceName);
+					return res.once(
+						cxt.status(200),
+						cxt.json({
 							namespace_id: "some-namespace-id",
 							namespace_name: "namespace-name",
 							created_on: "2022-06-29T14:30:08.16152Z",
 							created_by: "1fc1df98cc4420fe00367c3ab68c1639",
 							modified_on: "2022-06-29T14:30:08.16152Z",
 							modified_by: "1fc1df98cc4420fe00367c3ab68c1639",
-						},
-					];
+						})
+					);
 				}
-			);
-			return requests;
-		}
+			)
+		);
 
 		it("should list all namespaces", async () => {
-			const requests = mockListRequest();
 			await runWrangler("dispatch-namespace list");
-			expect(requests.count).toBe(1);
 			expect(std.out).toMatchInlineSnapshot(`
 			"[
 			  {
@@ -266,28 +256,30 @@ describe("dispatch-namespace", () => {
 	});
 
 	describe("rename namespace", () => {
-		function mockRenameRequest(expectedName: string) {
-			const requests = { count: 0 };
-			setMockResponse(
-				"/accounts/:accountId/workers/dispatch/namespaces/:namespaceName",
-				([_url, _, namespaceName], init) => {
-					requests.count += 1;
-
-					expect(init.method).toEqual("PUT");
-					expect(namespaceName).toEqual(expectedName);
-
-					return {
-						namespace_id: "some-namespace-id",
-						namespace_name: "namespace-name",
-						created_on: "2022-06-29T14:30:08.16152Z",
-						created_by: "1fc1df98cc4420fe00367c3ab68c1639",
-						modified_on: "2022-06-29T14:30:08.16152Z",
-						modified_by: "1fc1df98cc4420fe00367c3ab68c1639",
-					};
+		const namespaceName = "my-namespace";
+		let counter = 0;
+		msw.use(
+			rest.put(
+				"/accounts/:accountId/workers/dispatch/namespaces/:namespaceNameParam",
+				(req, res, cxt) => {
+					counter++;
+					const { namespaceNameParam } = req.params;
+					expect(counter).toBe(1);
+					expect(namespaceNameParam).toBe(namespaceName);
+					return res.once(
+						cxt.status(200),
+						cxt.json({
+							namespace_id: "some-namespace-id",
+							namespace_name: "namespace-name",
+							created_on: "2022-06-29T14:30:08.16152Z",
+							created_by: "1fc1df98cc4420fe00367c3ab68c1639",
+							modified_on: "2022-06-29T14:30:08.16152Z",
+							modified_by: "1fc1df98cc4420fe00367c3ab68c1639",
+						})
+					);
 				}
-			);
-			return requests;
-		}
+			)
+		);
 
 		it("should display help for rename", async () => {
 			await expect(
@@ -314,13 +306,11 @@ describe("dispatch-namespace", () => {
 		});
 
 		it("should attempt to rename the given namespace", async () => {
-			const namespaceName = "my-namespace";
 			const newName = "new-namespace";
-			const requests = mockRenameRequest(namespaceName);
 			await runWrangler(
 				`dispatch-namespace rename ${namespaceName} ${newName}`
 			);
-			expect(requests.count).toBe(1);
+
 			expect(std.out).toMatchInlineSnapshot(
 				`"Renamed dispatch namespace \\"my-namespace\\" to \\"new-namespace\\""`
 			);