+++ packages/wrangler/src/dev/local.tsx
@@ -25,6 +25,7 @@ import type {
 	CfKvNamespace,
 	CfR2Bucket,
 	CfVars,
+	CfD1Database,
 } from "../worker";
 import type { EsbuildBundle } from "./use-esbuild";
 import type {
@@ -59,7 +60,6 @@ export interface LocalProps {
 	localUpstream: string | undefined;
 	inspect: boolean;
 	onReady: ((ip: string, port: number) => void) | undefined;
-	logLevel: "none" | "error" | "log" | "warn" | "debug" | undefined;
 	logPrefix?: string;
 	enablePagesAssetsServiceBinding?: EnablePagesAssetsServiceBindingOptions;
 	testScheduled?: boolean;
@@ -101,7 +101,6 @@ function useLocalWorker({
 	localUpstream,
 	inspect,
 	onReady,
-	logLevel,
 	logPrefix,
 	enablePagesAssetsServiceBinding,
 	experimentalLocal,
@@ -188,6 +187,7 @@ function useLocalWorker({
 				usageModel,
 				kv_namespaces: bindings?.kv_namespaces,
 				r2_buckets: bindings?.r2_buckets,
+				d1_databases: bindings?.d1_databases,
 				internalDurableObjects,
 				externalDurableObjects,
 				localPersistencePath,
@@ -199,7 +199,6 @@ function useLocalWorker({
 				dataBlobBindings,
 				crons,
 				upstream,
-				logLevel,
 				logPrefix,
 				workerDefinitions,
 				enablePagesAssetsServiceBinding,
@@ -352,6 +351,7 @@ function useLocalWorker({
 		bindings.durable_objects,
 		bindings.kv_namespaces,
 		bindings.r2_buckets,
+		bindings.d1_databases,
 		bindings.vars,
 		bindings.services,
 		workerDefinitions,
@@ -369,7 +369,6 @@ function useLocalWorker({
 		localProtocol,
 		localUpstream,
 		inspect,
-		logLevel,
 		logPrefix,
 		onReady,
 		enablePagesAssetsServiceBinding,
@@ -473,6 +472,7 @@ interface SetupMiniflareOptionsProps {
 	usageModel: "bundled" | "unbound" | undefined;
 	kv_namespaces: CfKvNamespace[] | undefined;
 	r2_buckets: CfR2Bucket[] | undefined;
+	d1_databases: CfD1Database[] | undefined;
 	internalDurableObjects: CfDurableObject[];
 	externalDurableObjects: CfDurableObject[];
 	localPersistencePath: string | null;
@@ -484,7 +484,6 @@ interface SetupMiniflareOptionsProps {
 	dataBlobBindings: Record<string, string>;
 	crons: Config["triggers"]["crons"];
 	upstream: string | undefined;
-	logLevel: "none" | "error" | "log" | "warn" | "debug" | undefined;
 	logPrefix: string | undefined;
 	workerDefinitions: WorkerRegistry | undefined;
 	enablePagesAssetsServiceBinding?: EnablePagesAssetsServiceBindingOptions;
@@ -503,6 +502,7 @@ export function setupMiniflareOptions({
 	usageModel,
 	kv_namespaces,
 	r2_buckets,
+	d1_databases,
 	internalDurableObjects,
 	externalDurableObjects,
 	localPersistencePath,
@@ -514,7 +514,6 @@ export function setupMiniflareOptions({
 	dataBlobBindings,
 	crons,
 	upstream,
-	logLevel,
 	logPrefix,
 	workerDefinitions,
 	enablePagesAssetsServiceBinding,
@@ -522,7 +521,6 @@ export function setupMiniflareOptions({
 	miniflareCLIPath: string;
 	forkOptions: string[];
 } {
-	// TODO: This was already messy with the custom `disableLogs` and `logOptions`.
 	// It's now getting _really_ messy now with Pages ASSETS binding outside and the external Durable Objects inside.
 	const options = {
 		name: workerName,
@@ -573,12 +571,14 @@ export function setupMiniflareOptions({
 				})
 				.filter(([_, details]) => !!details)
 		),
+		d1Databases: d1_databases?.map((db) => db.binding),
 		...(localPersistencePath
 			? {
 					cachePersist: path.join(localPersistencePath, "cache"),
 					durableObjectsPersist: path.join(localPersistencePath, "do"),
 					kvPersist: path.join(localPersistencePath, "kv"),
 					r2Persist: path.join(localPersistencePath, "r2"),
+					d1Persist: path.join(localPersistencePath, "d1"),
 			  }
 			: {
 					// We mark these as true, so that they'll
@@ -610,7 +610,7 @@ export function setupMiniflareOptions({
 		logUnhandledRejections: true,
 		crons,
 		upstream,
-		disableLogs: logLevel === "none",
+		logLevel: logger.loggerLevel,
 		logOptions: logPrefix ? { prefix: logPrefix } : undefined,
 		enablePagesAssetsServiceBinding,
 	};